requirements.txt
# Core dependencies
Pillow>=9.0.0
piexif>=1.1.3
pyheif>=0.7.0
requests>=2.25.0

# Vision-Language Models
torch>=2.0.0
transformers>=4.30.0
accelerate>=1.0.0
safetensors>=0.3.0
huggingface-hub>=0.17.0

# Search Application
Flask>=2.3.0
Flask-CORS>=4.0.0
SQLAlchemy>=2.0.0
diskcache>=5.0.0
tqdm>=4.65.0

# Test dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=24.0.0
flake8>=7.0.0
pre-commit>=3.5.0

# Optional dependencies for notebooks
jupyterlab>=4.0.0
ipywidgets>=8.0.0
matplotlib>=3.0.0